cleans use cases implementation
