+++ src/main/java/com/ciklum/gitlum/controller/Controller.java
@@ -2,7 +2,7 @@ package com.ciklum.gitlum.controller;
 
 
 import com.ciklum.gitlum.domain.model.dto.RepoDTO;
-import com.ciklum.gitlum.domain.usecase.ProduceResult;
+import com.ciklum.gitlum.domain.usecase.BuildGitRepository;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +22,7 @@ import static com.ciklum.gitlum.config.Constants.CONTROLLER_BASE_URL;
 @Slf4j
 public class Controller {
 
-	private final ProduceResult produceResult;
+	private final BuildGitRepository buildGitRepository;
 
 	@ApiOperation(
 			value = "Serves a collection of Github repositories and their branches per the business requirements",
@@ -32,7 +32,7 @@ public class Controller {
 	@ResponseStatus(HttpStatus.OK)
 	public Flux<RepoDTO> getResult(@RequestBody final Request request) {
 		log.info("Querying repositories and branches for user {}", request.gitUser());
-		return produceResult.invoke(request.gitUser());
+		return buildGitRepository.invoke(request.gitUser());
 	}
 
 }
