+++ src/main/java/com/ciklum/gitlum/domain/usecase/GetBranches.java
@@ -1,17 +1,12 @@
 package com.ciklum.gitlum.domain.usecase;
 
 import com.ciklum.gitlum.annotations.UseCase;
-import com.ciklum.gitlum.domain.model.dto.BranchDTO;
 import com.ciklum.gitlum.domain.model.git.Branch;
 import com.ciklum.gitlum.domain.model.git.Repo;
-import com.ciklum.gitlum.domain.model.mappers.BranchMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.reactive.function.client.WebClient;
-import reactor.core.publisher.Mono;
-
-import java.util.Set;
-import java.util.stream.Collectors;
+import reactor.core.publisher.Flux;
 
 @UseCase
 @RequiredArgsConstructor
@@ -23,9 +18,8 @@ public class GetBranches {
 	private String repositoryBranchesURI;
 
 	private final WebClient.Builder webClientBuilder;
-	private final BranchMapper branchMapper;
 
-	public Mono<Set<BranchDTO>> invoke(final Repo source) {
+	public Flux<Branch> invoke(final Repo source) {
 		return webClientBuilder
 				.baseUrl(gitBaseURL)
 				.build()
@@ -36,8 +30,6 @@ public class GetBranches {
 								.build(source.getOwner().getLogin(), source.getName())
 				)
 				.retrieve()
-				.bodyToFlux(Branch.class)
-				.map(branchMapper::entityToDTO)
-				.collect(Collectors.toSet());
+				.bodyToFlux(Branch.class);
 	}
 }
